fix bugs

+++ utils.py
@@ -78,7 +78,7 @@ def parse_args():
                         help="path of directory containing tasks")
     parser.add_argument('--gui', action='store_true',
                         default=False, help='Run headless or render')
-    parser.add_argument('--max_time', type=int,
+    parser.add_argument('--max_time', type=float,
                         default=None, help='Maximum train run time in hours')
     parser.add_argument('--num_processes', type=int,
                         default=16, help='How many processes to parallelize')
